Simplify broker and test delete failure recursively
The broker was full of a bunch of stuff unrelated to its trusted mission:
clobbering un-whitelisted properties. So, we simplify, removing everything
unrelated to this mission.

It now accepts a config object with two properties:

- properties: what to do with the global properties
  - ignore: the whitelist of property names to not clobber
- scripts: an array of scripts to run in order. importScripts is clobbered
  beforehand, so all scripts to run must be specified when the broker is
  configured

Also refactor the tests to check the result of `delete`ing all the
un-whitelisted properties of the global object.